+++ apps/docs/src/examples/slider/multithrumb.tsx
@@ -1,0 +1,5 @@
+import { Slider } from "@pixelshades/ui/components"
+
+export default function Example() {
+	return <Slider label="Range" helperText="helper text" thumbLabels={["start", "end"]} defaultValue={[30, 60]} />
+}