+++ packages/ui/src/components/ui/date-range-picker/date-range-picker.tsx
@@ -4,10 +4,10 @@
 //
 // SPDX-License-Identifier: Apache-2.0
 
-import type { DatePickerProps as AriaDatePickerProps, ValidationResult } from "react-aria-components"
+import type { DateRangePickerProps as AriaDateRangePickerProps, ValidationResult } from "react-aria-components"
 import {
 	Button as AriaButton,
-	DatePicker as AriaDatePicker,
+	DateRangePicker as AriaDateRangePicker,
 	Dialog as AriaDialog,
 	DateInput,
 	DateSegment,
@@ -17,24 +17,23 @@ import {
 import { ChevronDownIcon } from "lucide-react"
 
 import { datePickerVariants } from "@pixelshades/styles/components/date-picker"
-import { inputVariants } from "@pixelshades/styles/components/input"
 import type { ReactNode } from "react"
 import { If } from "../../utils"
-import { Calendar } from "../calendar"
 import { FormDescription, FormFieldError, FormFieldGroup } from "../form"
 import { type FormComponentLabelProps, Label } from "../label"
 import { Popover } from "../popover"
+import { RangeCalendar } from "../range-calendar"
 
 const { inputButton, base, dateInput } = datePickerVariants()
 
-interface DatePickerProps<T extends DateValue> extends AriaDatePickerProps<T>, FormComponentLabelProps {
+interface DateRangePickerProps<T extends DateValue> extends AriaDateRangePickerProps<T>, FormComponentLabelProps {
 	/** A helper text to display below the date picker. */
 	helperText?: ReactNode
 	/** The error message to display when the date picker is invalid. */
 	errorMessage?: string | ((validation: ValidationResult) => string)
 }
 
-const DatePicker = <T extends DateValue>({
+const DateRangePicker = <T extends DateValue>({
 	className,
 	label,
 	description,
@@ -44,13 +43,17 @@ const DatePicker = <T extends DateValue>({
 	children,
 	isRequired,
 	...props
-}: DatePickerProps<T>) => (
-	<AriaDatePicker className="flex flex-col gap-md" isRequired={isRequired} {...props}>
+}: DateRangePickerProps<T>) => (
+	<AriaDateRangePicker className="flex flex-col gap-md" isRequired={isRequired} {...props}>
 		<Label description={description} tooltip={tooltip} isRequired={isRequired}>
 			{label}
 		</Label>
 		<FormFieldGroup className={"relative flex items-center"}>
-			<DateInput className={dateInput()}>
+			<DateInput className={dateInput()} slot="start">
+				{(segment) => <DateSegment className={base()} segment={segment} />}
+			</DateInput>
+			<span aria-hidden="true">–</span>
+			<DateInput className={dateInput()} slot="end">
 				{(segment) => <DateSegment className={base()} segment={segment} />}
 			</DateInput>
 			<AriaButton className={inputButton()}>
@@ -63,10 +66,10 @@ const DatePicker = <T extends DateValue>({
 		<FormFieldError>{errorMessage}</FormFieldError>
 		<Popover className="w-fit border-none p-0">
 			<AriaDialog>
-				<Calendar />
+				<RangeCalendar />
 			</AriaDialog>
 		</Popover>
-	</AriaDatePicker>
+	</AriaDateRangePicker>
 )
 
-export { type DatePickerProps, DatePicker }
+export { type DateRangePickerProps, DateRangePicker }